servicioPubSubConversor/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==5.5
Deprecated==1.2.13
Faker==14.0.0
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.2.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
packaging==21.3
prompt-toolkit==3.0.31
PyJWT==2.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.15
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.14.1
requests==2.28.1
Flask==1.1.2
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-pubsub==2.13.11
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
psycopg2-binary==2.9.5
